Tidy up go-online/go-offline
Also icons are updated with ++LITE for lite version
Also progress spinner is factored out into a class and simpler to use
Also some issues with loading db if it's empty or the root node is not present are fixed
Also some issues with not recreating deleted account/settings etc folders after cache wipe
Also making sure callbacks from MEGA are marshalled to the main gui thread for processing
Also make sure we request download of playlists that are already visible after cache wipe, and report absence nicely

+++ src/StorageModel.swift
@@ -164,6 +164,13 @@ class StorageModel {
         catch {
             result = false;
         }
+        
+        // recreate those folders again (now empty) so we don't have issues with the next login etc.
+        app().storageModel.alreadyCreatedFolders = [];
+        _ = accountPath();
+        _ =  cacheFilesPath();
+        _ = tempFilesPath();
+        
         return result;
     }
 
@@ -477,6 +484,7 @@ class StorageModel {
     {
         do {
             try content.write(toFile: settingsPath() + "/" + leafname, atomically: true, encoding: String.Encoding.utf8);
+            print("Wrote file: " + settingsPath() + "/" + leafname)
             return true;
         }
         catch {
@@ -486,11 +494,14 @@ class StorageModel {
     
     func deleteSettingFile(leafname : String)
     {
-        do {
-            try FileManager.default.removeItem(atPath: settingsPath() + "/" + leafname)
-        }
-        catch {
-            print("Failed to remove file " + settingsPath() + "/" + leafname)
+        if FileManager.default.fileExists(atPath: settingsPath() + "/" + leafname) {
+            do {
+                try FileManager.default.removeItem(atPath: settingsPath() + "/" + leafname)
+                print("Removed file: " + settingsPath() + "/" + leafname)
+            }
+            catch {
+                print("Failed to remove file " + settingsPath() + "/" + leafname)
+            }
         }
     }
